requirements.txt
blinker==1.9.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
gunicorn==23.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
rpi.gpio==0.7.1
sqlalchemy==2.0.38; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
